move method to shell
