+++ src/gluonts/model/forecast.py
@@ -188,32 +188,7 @@ class Forecast:
         if OutputType.samples in config.output_types:
             result["samples"] = []
 
-        return Forecast.jsonify_floats(result)
-
-    @staticmethod
-    def jsonify_floats(json_object):
-        """
-        Traverses through the JSON object and converts non JSON-spec compliant
-        floats(nan, -inf, inf) to their string representations.
-
-        Parameters
-        ----------
-        json_object
-            JSON object
-        """
-        if isinstance(json_object, dict):
-            return {k: Forecast.jsonify_floats(v) for k, v in json_object.items()}
-        elif isinstance(json_object, list):
-            return [Forecast.jsonify_floats(item) for item in json_object]
-        elif isinstance(json_object, float):
-            if np.isnan(json_object):
-                return "NaN"
-            elif np.isposinf(json_object):
-                return "Infinity"
-            elif np.isneginf(json_object):
-                return "-Infinity"
-            return json_object
-        return json_object
+        return result
 
 
 class SampleForecast(Forecast):
@@ -343,7 +318,7 @@ class SampleForecast(Forecast):
         result = super().as_json_dict(config)
 
         if OutputType.samples in config.output_types:
-            result["samples"] = Forecast.jsonify_floats(self.samples.tolist())
+            result["samples"] = self.samples.tolist()
 
         return result
